requirements.txt
flask
requests
pdfplumber
python-docx
beautifulsoup4
python-slugify
playwright
lxml
pillow
reportlab